scripts/print_arm_pose/requirements.txt
-f ../../../prebuilt
bosdyn-client >= 3.1
cv2 >= 4.8.1